fastboot: Boot properly from USB mfgtool usage.
This gets fastboot to flash to MMC0 properly when used as the bootloader from
mfgtool.

NOTE: To get the board to boot from the newly flashed u-boot, we must also issue
the "mmc partconf 0 0 1 0" command to make the eMMC boot flag point to the first
partition and cause the SPL ROM bootloader to chainload from the correct
location. Failure to do so will result in the board continuing to enter serial
download mode.

Change-Id: Idba951ceaee21990020a908c8b65bd6df5ae383f